Create a nice base class for the needed structures!
