+++ tests/active_learning_loop/excisor/base_test_excision.py
@@ -2,9 +2,10 @@ import numpy as np
 import pytest
 
 from diffusion_for_multi_scale_molecular_dynamics.namespace import AXL
+from tests.active_learning_loop.conftest import BaseTestAxlStructure
 
 
-class BaseTestExcision:
+class BaseTestExcision(BaseTestAxlStructure):
 
     @pytest.fixture(scope="class", autouse=True)
     def set_random_seed(self):
@@ -21,60 +22,6 @@ class BaseTestExcision:
     @pytest.fixture
     def expected_excised_atom_index(self, **kwargs):
         raise NotImplementedError("must be implemented in subclass")
-
-    @pytest.fixture
-    def number_of_atoms(self):
-        return 48
-
-    @pytest.fixture(params=[1, 2, 3])
-    def spatial_dimension(self, request):
-        return request.param
-
-    @pytest.fixture
-    def basis_vectors(self, spatial_dimension):
-        box_size = np.random.random((spatial_dimension,))
-        return np.diag(box_size)
-
-    @pytest.fixture
-    def lattice_parameters(self, spatial_dimension, basis_vectors):
-        lp = np.concatenate(
-            (
-                np.diag(basis_vectors),
-                np.zeros(int(spatial_dimension * (spatial_dimension - 1) / 2)),
-            )
-        )
-        return lp
-
-    @pytest.fixture
-    def atom_relative_coordinates(self, number_of_atoms, spatial_dimension):
-        return np.random.random((number_of_atoms, spatial_dimension))
-
-    @pytest.fixture
-    def atom_cartesian_positions(self, atom_relative_coordinates, basis_vectors):
-        return np.matmul(atom_relative_coordinates, basis_vectors)
-
-    @pytest.fixture
-    def num_atom_types(self):
-        return 6
-
-    @pytest.fixture
-    def atom_species(self, num_atom_types, number_of_atoms):
-        return np.random.randint(low=0, high=num_atom_types, size=(number_of_atoms,))
-
-    @pytest.fixture
-    def structure_axl(self, atom_species, atom_relative_coordinates, lattice_parameters):
-        struct_axl = AXL(
-            A=atom_species, X=atom_relative_coordinates, L=lattice_parameters
-        )
-        return struct_axl
-
-    @pytest.fixture
-    def atom_uncertainty(self, number_of_atoms):
-        return np.random.random((number_of_atoms,))
-
-    @pytest.fixture(params=range(48))
-    def central_atom_idx(self, request):
-        return request.param
 
     def test_excise_one_environment(
         self,
